[F] issue #2
https://github.com/remobile/react-native-refresh-infinite-listview/issues/2

When the list is not scrolled to the top, then scroll down until beyond top, the pull down to refresh indicator is not showing.

And also:

When the list is not scrolled to the bottom, then scroll up until beyond bottom, the pull up to load more indicator is not showing.